Preparations for using remote posterior databases in python
This moves all filesystem related logic to the PosteriorDatabase class. Later we
can add a CachingRemotePosteriorDatabase that downloads the files from the
remote posterior database (for example a github repo) and caches them locally.
Then we can add a convenience function that creates a local posterior db if
passed a local filepath and a remote posterior db if passed an url, and if
passed nothing it can default to our github repository.