python: Remove dispatch methods from Posterior
This removes the methods from Posterior that dispatch to methods of Model or
Data.

So now for example `posterior.model_code("stan")` is written
`posterior.model.code("stan")`, so the new code calls the `code` method on `Model`
instead of `model_code` on `Posterior`.

The aim of this change is to make the library more simple. This change aligns
well with the Zen of Python guideline "There should be one-- and preferably only one
--obvious way to do it.".

Of course if having two ways of doing something adds value then it can be worth
the duplication, however in this case I don't really see a big difference
between `posterior.model_code("stan")` and `posterior.model.code("stan")`.