Test now works locally
Or at least sort of works. Because the readme contains absolute paths (for
example the result of `dfp <- data_file_path(po)`) the generated result depends
on who runs it.

Currently the test also requires that for example the `posterior` package is
installed locally, which won't be the case in CI currently.